moved auth controller code to model, created account model & controller
